Presently, basic tests are passing.  The system is basically reflective, but does no CRUD work.
I had changed the system so that components are created centrally, rather than
by their parents.  I think now that was a mistake.  It seemed like a
simplification at the time, but I think I was focused too much on the analysis
phase, which will be where a relatively small amount of work is done.  I think
now that letting each component handle its own business, and doing the analysis
recursively, was better after all.  Accordingly, I am changing back.

I have also thought through how to share database access, and I have a better
idea of how I want to handle it.  This is part of the decision to re-decentralize.
(Cool word, huh?)  It may be that I'm still not handling the shared resource in
the best way, but I think that any better way will still support a
decentralized system, and that the change will not be too painful.