In accordance with prophecy, I have removed most of the functionality from dbo.rb, and am teaching database.rb to do its own work.  Not quite sure that's solid, yet.  After that, I'll do the lower level classes.
Passing all tests, but the tests are a bit messy.  I need to clean them up, so
that I'm sure they tell me what I want.

I need to establish a mechanism for sharing the correct database connection
amongst the objects which need it.  Making it available to their classes isn't
going to work, since there can be more than one database connection and a given
instance of Schema, Table, etc. belongs to only one.

I've gone a bit too "meta" with this.  As part of getting it started, for
instance, I've been auto-generating instance variables and accessors based on
query results, and then assigning aliases.  That was fine for getting started,
though I probably went a little far with it for that purpose.  But the results
from information_schema should be the same everywhere, so there's no real
reason to auto-assign those.  Also, I know the names I want to use from the
outset, so aliases aren't needed.  I'm going to change things to declare
accessors conventionally, and just assign each instance variable explicitly in
the constructor.