The DBO Ruby library provides DataBase Objects.  This is the initial commit of a rewrite of the original, unreleased vesion.
At present, it has a templating system, a database profile manager,
and it can establish database connections to PostgreSQL.  Most of the
actual Objects, representing the different entities in a database, are
not properly implemented.  Classes like DBO::Table and DBO::Column
are mere holdovers from the original, and need to be rewritten.

There is a decent starting suite of tests, all of which presently pass.
The tests offer a decent list of those parts of the project which are
up to date.